requirements.txt
argcomplete @ file:///home/conda/feedstock_root/build_artifacts/argcomplete_1619128689661/work
argon2-cffi @ file:///home/conda/feedstock_root/build_artifacts/argon2-cffi_1610522577486/work
async-generator==1.10
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1620387926260/work
backcall @ file:///home/conda/feedstock_root/build_artifacts/backcall_1592338393461/work
backports.functools-lru-cache @ file:///home/conda/feedstock_root/build_artifacts/backports.functools_lru_cache_1618230623929/work
bleach @ file:///home/conda/feedstock_root/build_artifacts/bleach_1629908509068/work
brotlipy==0.7.0
cached-property==1.5.2
certifi==2021.5.30
cffi @ file:///tmp/build/80754af9/cffi_1625814693446/work
cftime==1.5.0
chardet @ file:///tmp/build/80754af9/chardet_1607706768982/work
colorama==0.4.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262151086/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616769182610/work
cycler==0.10.0
debugpy @ file:///home/conda/feedstock_root/build_artifacts/debugpy_1627074840854/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1621187651333/work
defusedxml @ file:///home/conda/feedstock_root/build_artifacts/defusedxml_1615232257335/work
entrypoints @ file:///home/conda/feedstock_root/build_artifacts/entrypoints_1605121927639/work/dist/entrypoints-0.3-py2.py3-none-any.whl
h5py==3.3.0
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
imageio==2.9.0
importlib-metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1630014178064/work
ipykernel @ file:///home/conda/feedstock_root/build_artifacts/ipykernel_1629116249491/work/dist/ipykernel-6.2.0-py3-none-any.whl
ipython @ file:///home/conda/feedstock_root/build_artifacts/ipython_1627911644410/work
ipython-genutils==0.2.0
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1610481889018/work
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1610146791823/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1621419064915/work
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1614815863336/work
jupyter==1.0.0
jupyter-client @ file:///tmp/build/80754af9/jupyter_client_1616770841739/work
jupyter-console @ file:///tmp/build/80754af9/jupyter_console_1616615302928/work
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions @ file:///home/conda/feedstock_root/build_artifacts/jupyter_contrib_nbextensions_1614931162960/work
jupyter-core @ file:///home/conda/feedstock_root/build_artifacts/jupyter_core_1612125257337/work
jupyter-highlight-selected-word @ file:///home/conda/feedstock_root/build_artifacts/jupyter_highlight_selected_word_1611341004115/work
jupyter-latex-envs @ file:///home/conda/feedstock_root/build_artifacts/jupyter_latex_envs_1614852190293/work
jupyter-nbextensions-configurator @ file:///home/conda/feedstock_root/build_artifacts/jupyter_nbextensions_configurator_1611341112910/work
jupyterlab-pygments @ file:///home/conda/feedstock_root/build_artifacts/jupyterlab_pygments_1601375948261/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
kiwisolver==1.3.1
lxml @ file:///home/conda/feedstock_root/build_artifacts/lxml_1616532288511/work
MarkupSafe @ file:///home/conda/feedstock_root/build_artifacts/markupsafe_1621455667184/work
matplotlib==3.4.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1618935594181/work
meshio==5.0.0
mistune @ file:///home/conda/feedstock_root/build_artifacts/mistune_1624941291602/work
mkl-fft==1.3.0
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
nbclient @ file:///home/conda/feedstock_root/build_artifacts/nbclient_1629120697898/work
nbconvert @ file:///home/conda/feedstock_root/build_artifacts/nbconvert_1624472657203/work
nbformat @ file:///home/conda/feedstock_root/build_artifacts/nbformat_1617383142101/work
nest-asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1617163391303/work
netCDF4==1.5.7
networkx==2.6.2
nibabel==3.2.1
notebook @ file:///home/conda/feedstock_root/build_artifacts/notebook_1628690102344/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1626264701743/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1625323647219/work
pandocfilters==1.4.3
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1617148930513/work
pexpect @ file:///home/conda/feedstock_root/build_artifacts/pexpect_1602535608087/work
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
Pillow==8.3.1
prometheus-client @ file:///home/conda/feedstock_root/build_artifacts/prometheus_client_1622586138406/work
prompt-toolkit==3.0.19
ptyprocess @ file:///home/conda/feedstock_root/build_artifacts/ptyprocess_1609419310487/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pycollada==0.7.1
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1629119114968/work
PyMCubes==0.1.2
pymesh2 @ file:///mnt/c/Users/Quinten/Projects/Iso%20surfaces/pymesh2-0.3-cp37-cp37m-linux_x86_64.whl
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.2.1
qtconsole==5.1.1
QtPy @ file:///tmp/build/80754af9/qtpy_1629397026935/work
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016701961/work
scikit-image==0.18.2
scipy==1.7.1
Send2Trash @ file:///home/conda/feedstock_root/build_artifacts/send2trash_1628511208346/work
six @ file:///tmp/build/80754af9/six_1623709665295/work
terminado @ file:///home/conda/feedstock_root/build_artifacts/terminado_1629327958745/work
testpath @ file:///home/conda/feedstock_root/build_artifacts/testpath_1621261527237/work
tifffile==2021.8.8
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1610094704591/work
tqdm @ file:///tmp/build/80754af9/tqdm_1625563689033/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1602771532708/work
typing-extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1622748266870/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1625084269274/work
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1600965781394/work
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1625284368454/work
